Refactor thomas model results to separate component
